fix: Clip.notes_countをi64+specta::Numberアノテーションに戻す

+++ src-tauri/src/api/clips.rs
@@ -16,7 +16,7 @@ struct RawClip {
     #[serde(default)]
     description: Option<String>,
     is_public: bool,
-    notes_count: u32,
+    notes_count: i64,
 }
 
 impl From<RawClip> for Clip {
