+++ src-tauri/src/domain/clip.rs
@@ -10,5 +10,6 @@ pub struct Clip {
     pub name: String,
     pub description: Option<String>,
     pub is_public: bool,
-    pub notes_count: u32,
+    #[specta(type = specta_typescript::Number)]
+    pub notes_count: i64,
 }
